+++ GRDB/Core/Row.swift
@@ -323,7 +323,7 @@ extension Row {
     /// the same name, the leftmost column is considered.
     ///
     /// The result is nil if the row does not contain the column.
-    public func value(column: Column) -> DatabaseValueConvertible? {
+    public func value(_ column: Column) -> DatabaseValueConvertible? {
         return value(named: column.name)
     }
     
@@ -335,7 +335,7 @@ extension Row {
     /// If the column is missing or if the SQLite value is NULL, the result is
     /// nil. Otherwise the SQLite value is converted to the requested type
     /// `Value`. Should this conversion fail, a fatal error is raised.
-    public func value<Value: DatabaseValueConvertible>(column: Column) -> Value? {
+    public func value<Value: DatabaseValueConvertible>(_ column: Column) -> Value? {
         return value(named: column.name)
     }
     
@@ -351,7 +351,7 @@ extension Row {
     /// This method exists as an optimization opportunity for types that adopt
     /// StatementColumnConvertible. It *may* trigger SQLite built-in conversions
     /// (see https://www.sqlite.org/datatype3.html).
-    public func value<Value: DatabaseValueConvertible & StatementColumnConvertible>(column: Column) -> Value? {
+    public func value<Value: DatabaseValueConvertible & StatementColumnConvertible>(_ column: Column) -> Value? {
         return value(named: column.name)
     }
     
@@ -364,7 +364,7 @@ extension Row {
     ///
     /// This method crashes if the fetched SQLite value is NULL, or if the
     /// SQLite value can not be converted to `Value`.
-    public func value<Value: DatabaseValueConvertible>(column: Column) -> Value {
+    public func value<Value: DatabaseValueConvertible>(_ column: Column) -> Value {
         return value(named: column.name)
     }
     
@@ -381,7 +381,7 @@ extension Row {
     /// This method exists as an optimization opportunity for types that adopt
     /// StatementColumnConvertible. It *may* trigger SQLite built-in conversions
     /// (see https://www.sqlite.org/datatype3.html).
-    public func value<Value: DatabaseValueConvertible & StatementColumnConvertible>(column: Column) -> Value {
+    public func value<Value: DatabaseValueConvertible & StatementColumnConvertible>(_ column: Column) -> Value {
         return value(named: column.name)
     }
     
@@ -429,7 +429,7 @@ extension Row {
     ///
     /// The returned data does not owns its bytes: it must not be used longer
     /// than the row's lifetime.
-    public func dataNoCopy(column: Column) -> Data? {
+    public func dataNoCopy(_ column: Column) -> Data? {
         return dataNoCopy(named: column.name)
     }
 }