+++ Tests/Public/QueryInterfaceRequest/QueryInterfaceRequestTests.swift
@@ -425,7 +425,7 @@ class QueryInterfaceRequestTests: GRDBTestCase {
         let dbQueue = try! makeDatabaseQueue()
         XCTAssertEqual(
             sql(dbQueue, tableRequest.reversed()),
-            "SELECT * FROM \"readers\" ORDER BY \"_rowid_\" DESC")
+            "SELECT * FROM \"readers\" ORDER BY \"rowid\" DESC")
         XCTAssertEqual(
             sql(dbQueue, tableRequest.order(Col.age).reversed()),
             "SELECT * FROM \"readers\" ORDER BY \"age\" DESC")